Rename the gripper dummy
That's now more in line with the other new packages.
