--- schunk_egu_egk_gripper_dummy/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/schunk_egu_egk_gripper_dummy
-[install]
-install_scripts=$base/lib/schunk_egu_egk_gripper_dummy
