+++ schunk_gripper_dummy/config/README.md
