+++ schunk_gripper_dummy/config/__init__.py
