+++ schunk_gripper_dummy/config/data.json
