+++ schunk_gripper_dummy/config/enum.json
