+++ schunk_gripper_dummy/config/metadata.json
