+++ schunk_gripper_dummy/config/read_system_parameters.py
