+++ schunk_gripper_dummy/doc/plc_input_data_frame.png
