+++ schunk_gripper_dummy/doc/plc_output_data_frame.png
