+++ schunk_gripper_dummy/schunk_gripper_dummy/main.py
