+++ schunk_gripper_dummy/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/schunk_gripper_dummy
+[install]
+install_scripts=$base/lib/schunk_gripper_dummy
