+++ schunk_gripper_dummy/src/dummy.py
