+++ schunk_gripper_dummy/tests/test_dummy.py
