+++ schunk_gripper_dummy/tests/test_plc_communication.py
